+++ src/server/services/pullRequest.js
@@ -13,7 +13,7 @@ const commentText = (signed, badgeUrl, claUrl, userMap, recheckUrl) => {
     }
 
     let youAll = (committersCount > 1 ? 'you all' : 'you')
-    let text = `[![CLA assistant check](${badgeUrl})](${claUrl}) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that ${youAll} sign our [Contributor License Agreement](${claUrl}) before we can accept your contribution.<br/>`
+    let text = `[![CLA assistant check](${badgeUrl})](${claUrl}) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that ${youAll} sign our [Contributor License Agreement](${claUrl}) before we can accept your contribution.<br/>`
     if (committersCount > 1) {
         text += '**' + userMap.signed.length + '** out of **' + (userMap.signed.length + userMap.not_signed.length) + '** committers have signed the CLA.<br/>'
         userMap.signed.forEach(function (signee) {
@@ -41,6 +41,11 @@ const commentText = (signed, badgeUrl, claUrl, userMap, recheckUrl) => {
 class PullRequestService {
     async badgeComment(owner, repo, pullNumber, signed, userMap) {
         let badgeUrl = url.pullRequestBadge(signed)
+        let fun
+        const arg = {
+            owner: owner,
+            repo: repo
+        }
         try {
             const comment = await this.getComment({
                 repo: repo,
@@ -50,30 +55,35 @@ class PullRequestService {
 
             const claUrl = url.claURL(owner, repo, pullNumber)
             const recheckUrl = url.recheckPrUrl(owner, repo, pullNumber)
-            const body = commentText(signed, badgeUrl, claUrl, userMap, recheckUrl)
+            arg.body = commentText(signed, badgeUrl, claUrl, userMap, recheckUrl)
 
-            let fun
-            const arg = {
-                owner: owner,
-                repo: repo,
-                body: body,
-                noCache: true
-            }
+
             if (!comment && !signed) {
                 fun = 'createComment'
                 arg.issue_number = pullNumber
             } else if (comment && comment.id) {
+                //Temporary fix of comments from outdated user
+                if (comment.user && comment.user.login === 'claassistantio' && config.server.github.token_old) {
+                    arg.comment_id = comment.id
+                    return updateCommentOfDeprecatedUser(arg, pullNumber, owner, repo)
+                } else if (arg.body === comment.body) {
+                    logger.debug(`Skip updateComment for the PR ${url.githubHttpPullRequest(owner, repo, pullNumber)} as there are no text changes`)
+                    return
+                }
                 fun = 'updateComment'
                 arg.comment_id = comment.id
             } else {
                 return
             }
 
-            await github.call({
+            github.call({
                 obj: 'issues',
                 fun,
                 arg,
                 token: config.server.github.token,
+            }).catch((error) => {
+                logger.debug(`Failed on api call issues/${fun} for PR ${url.githubHttpPullRequest(owner, repo, pullNumber)}`)
+                logger.warn(new Error(error).stack)
             })
         } catch (error) {
             logger.warn(new Error(error).stack)
@@ -87,8 +97,7 @@ class PullRequestService {
             arg: {
                 owner: args.owner,
                 repo: args.repo,
-                issue_number: args.number,
-                noCache: true
+                issue_number: args.number
             },
             token: config.server.github.token
         })
@@ -120,8 +129,7 @@ class PullRequestService {
                     owner: args.owner,
                     repo: args.repo,
                     comment_id: comment.id,
-                    body: body,
-                    noCache: true
+                    body: body
                 },
                 token: config.server.github.token
             })
@@ -158,3 +166,25 @@ class PullRequestService {
 
 
 module.exports = new PullRequestService()
+
+//Temporary fix of comments from outdated user -- remove later
+function updateCommentOfDeprecatedUser(arg, pullNumber, owner, repo) {
+    github.call({
+        obj: 'issues',
+        fun: 'deleteComment',
+        arg,
+        token: config.server.github.token_old,
+    }).catch(() => {
+        logger.debug('Failed on deleting comment from the old user')
+    })
+    arg.issue_number = pullNumber
+    github.call({
+        obj: 'issues',
+        fun: 'createComment',
+        arg,
+        token: config.server.github.token,
+    }).catch(() => {
+        logger.debug('Failed on creating comment for CLAassistant user')
+    })
+    logger.debug(`Changing comment user for PR ${url.githubHttpPullRequest(owner, repo, pullNumber)}`)
+}